+++ src/index.ts
@@ -1,10 +1,14 @@
 export {
   createProxy,
+  createReverseProxy,
   type ProxyOptions,
   type ProxyProfile,
+  type ReverseProxyOptions,
   type RequestLog,
   type RequestOutcome,
 } from "./proxy.js";
+export { type BlackoutConfig, isInBlackout } from "./levers/blackout.js";
 export { type LatencyConfig, sampleLatencyMs } from "./levers/latency.js";
 export { type LossConfig, shouldDropConnection } from "./levers/loss.js";
 export { PROFILES, PROFILE_NAMES, type ProfileName, getProfile } from "./profiles.js";
+export { deriveConventionalPort } from "./ports.js";